Unify bucket creation and cleanup
- Remove duplicated logic
- Use a world-managed function to keep track of created objects
- Clean the buckets at the end of any test, unless it failed, to
  ease debugging.

Issue: ZENKO-4898

+++ tests/ctst/common/common.ts
@@ -31,25 +31,33 @@ export async function cleanS3Bucket(
     if (!bucketName) {
         return;
     }
-    world.resetCommand();
-    world.addCommandParameter({ bucket: bucketName });
-    const createdObjects = world.getSaved<Map<string, string>>('createdObjects');
-    if (createdObjects !== undefined) {
-        const results = await S3.listObjectVersions(world.getCommandParameters());
-        const res = safeJsonParse<ListObjectVersionsOutput>(results.stdout);
-        assert(res.ok);
-        const versions = res.result!.Versions || [];
-        const deleteMarkers = res.result!.DeleteMarkers || [];
-        await Promise.all(versions.concat(deleteMarkers).map(obj => {
-            world.addCommandParameter({ key: obj.Key });
-            world.addCommandParameter({ versionId: obj.VersionId });
-            return S3.deleteObject(world.getCommandParameters());
-        }));
-        world.deleteKeyFromCommand('key');
-        world.deleteKeyFromCommand('versionId');
+    try {
+        Identity.useIdentity(IdentityEnum.ACCOUNT, world.getSaved<string>('accountName') ||
+            world.parameters.AccountName);
+        world.resetCommand();
+        world.addCommandParameter({ bucket: bucketName });
+        const createdObjects = world.getCreatedObjects();
+        if (createdObjects !== undefined) {
+            const results = await S3.listObjectVersions(world.getCommandParameters());
+            const res = safeJsonParse<ListObjectVersionsOutput>(results.stdout);
+            if (!res.ok) {
+                throw results;
+            }
+            const versions = res.result!.Versions || [];
+            const deleteMarkers = res.result!.DeleteMarkers || [];
+            await Promise.all(versions.concat(deleteMarkers).map(obj => {
+                world.addCommandParameter({ key: obj.Key });
+                world.addCommandParameter({ versionId: obj.VersionId });
+                return S3.deleteObject(world.getCommandParameters());
+            }));
+            world.deleteKeyFromCommand('key');
+            world.deleteKeyFromCommand('versionId');
+        }
+        await S3.deleteBucketLifecycle(world.getCommandParameters());
+        await S3.deleteBucket(world.getCommandParameters());
+    } catch (err) {
+        world.logger.warn('Error cleaning bucket', { bucketName, err });
     }
-    await S3.deleteBucketLifecycle(world.getCommandParameters());
-    await S3.deleteBucket(world.getCommandParameters());
 }
 
 async function addMultipleObjects(this: Zenko, numberObjects: number,
@@ -65,12 +73,7 @@ async function addMultipleObjects(this: Zenko, numberObjects: number,
         if (userMD) {
             this.addCommandParameter({ metadata: JSON.stringify(userMD) });
         }
-        this.addToSaved('objectName', objectNameFinal);
-        this.logger.debug('Adding object', { objectName: objectNameFinal });
         lastResult = await putObject(this, objectNameFinal);
-        const createdObjects = this.getSaved<Map<string, string>>('createdObjects') || new Map<string, string>();
-        createdObjects.set(this.getSaved<string>('objectName'), this.getSaved<string>('versionId'));
-        this.addToSaved('createdObjects', createdObjects);
     }
     return lastResult;
 }
@@ -154,7 +157,7 @@ Given('a tag on object {string} with key {string} and value {string}',
         this.resetCommand();
         this.addCommandParameter({ bucket: this.getSaved<string>('bucketName') });
         this.addCommandParameter({ key: objectName });
-        const versionId = this.getSaved<Map<string, string>>('createdObjects')?.get(objectName);
+        const versionId = this.getLatestObjectVersion(objectName);
         if (versionId) {
             this.addCommandParameter({ versionId });
         }
@@ -173,7 +176,7 @@ Then('object {string} should have the tag {string} with value {string}',
         this.resetCommand();
         this.addCommandParameter({ bucket: this.getSaved<string>('bucketName') });
         this.addCommandParameter({ key: objectName });
-        const versionId = this.getSaved<Map<string, string>>('createdObjects')?.get(objectName);
+        const versionId = this.getLatestObjectVersion(objectName);
         if (versionId) {
             this.addCommandParameter({ versionId });
         }
@@ -188,7 +191,7 @@ Then('object {string} should have the user metadata with key {string} and value 
         this.resetCommand();
         this.addCommandParameter({ bucket: this.getSaved<string>('bucketName') });
         this.addCommandParameter({ key: objectName });
-        const versionId = this.getSaved<Map<string, string>>('createdObjects')?.get(objectName);
+        const versionId = this.getLatestObjectVersion(objectName);
         if (versionId) {
             this.addCommandParameter({ versionId });
         }
@@ -220,7 +223,7 @@ When('i delete object {string}', async function (this: Zenko, objectName: string
     this.resetCommand();
     this.addCommandParameter({ bucket: this.getSaved<string>('bucketName') });
     this.addCommandParameter({ key: objName });
-    const versionId = this.getSaved<Map<string, string>>('createdObjects')?.get(objName);
+    const versionId = this.getLatestObjectVersion(objName);
     if (versionId) {
         this.addCommandParameter({ versionId });
     }
@@ -378,12 +381,7 @@ Given('an upload size of {int} B for the object {string}', async function (
 ) {
     this.addToSaved('objectSize', size);
     if (this.getSaved<boolean>('preExistingObject')) {
-        if (objectName) {
-            this.addToSaved('objectName', objectName);
-        } else {
-            this.addToSaved('objectName', `object-${Utils.randomString()}`);
-        }
-        await putObject(this, this.getSaved<string>('objectName'));
+        await putObject(this, objectName);
     }
 });
 
@@ -391,8 +389,7 @@ When('I PUT an object with size {int}', async function (this: Zenko, size: numbe
     if (size > 0) {
         this.addToSaved('objectSize', size);
     }
-    this.addToSaved('objectName', `object-${Utils.randomString()}`);
     const result = await addMultipleObjects.call(
-        this, 1, this.getSaved<string>('objectName'), size);
+        this, 1, `object-${Utils.randomString()}`, size);
     this.setResult(result!);
 });
